Implement a better interface to follow_links function.
The follow_links function accepts a test function as  an argument and uses it to test the obtained links. A link could either match the pattern specified by the test function and thus added to the matching_urls list. Or else, it would be added to the other_urls list.